Unify memory resource
