+++ config/template.yaml
@@ -18,7 +18,7 @@ plugins:
       base_url: "${OLLAMA_BASE_URL}"
       model: "${OLLAMA_MODEL}"
     memory:
-      type: pipeline.resources.memory_resource:MemoryResource
+      type: pipeline.resources.memory:Memory
       database:
         type: plugins.builtin.resources.sqlite_storage:SQLiteStorageResource
         path: ./agent.db
