--- src/plugins/builtin/resources/memory_resource.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-"""Expose memory resource implementations for plugin imports."""
-
-from pipeline.resources.memory_resource import MemoryResource, SimpleMemoryResource
-
-__all__ = ["MemoryResource", "SimpleMemoryResource"]
