+++ user_plugins/prompts/conversation_history.py
@@ -4,7 +4,7 @@ from typing import Dict, List
 
 from pipeline.base_plugins import PromptPlugin
 from pipeline.context import ConversationEntry, PluginContext
-from pipeline.resources.memory_resource import MemoryResource
+from pipeline.resources.memory import Memory
 from pipeline.stages import PipelineStage
 
 
@@ -18,7 +18,7 @@ class ConversationHistory(PromptPlugin):
         super().__init__(config)
 
     async def _execute_impl(self, context: PluginContext) -> None:
-        memory: MemoryResource = context.get_resource("memory")
+        memory: Memory = context.get_resource("memory")
         if memory is None:
             return
         if context.current_stage == PipelineStage.PARSE:
